ws server is able to receive and send messages
