+++ backend/src/ws/index.ts
@@ -1,1 +1,3 @@
 export { default as WSConnection } from './Connection';
+export { default as ValidateRequest } from './ValidateRequest';
+export { default as NotificationHandler } from './NotificationHandler';
